hf_updated_files/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
numpy<2.0.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
edge-tts==6.1.9
ffmpeg-python==0.2.0
